requirements.txt
sounddevice==0.4.7
soundfile==0.12.1
flask==3.0.3
google-generativeai==0.7.2
numpy==1.26.4
